refactor: modularize map component for improved maintainability
