+++ web/src/lib/utils.ts
@@ -7,3 +7,18 @@ import { twMerge } from "tailwind-merge"
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
 }
+
+/**
+ * HTMLエスケープを行うユーティリティ関数
+ * XSS攻撃を防ぐために特殊文字をエスケープします
+ */
+export function escapeHtml(text: string | number): string {
+  const escapeMap: { [key: string]: string } = {
+    '&': '&amp;',
+    '<': '&lt;',
+    '>': '&gt;',
+    '"': '&quot;',
+    "'": '&#039;'
+  };
+  return String(text).replace(/[&<>"']/g, (match) => escapeMap[match]);
+}
